Create a backend for the modal
